+++ src/test/java/backend/wal/home/domain/entity/HomeTest.java
@@ -2,8 +2,8 @@ package backend.wal.home.domain.entity;
 
 import backend.wal.home.domain.Home;
 import backend.wal.home.domain.OpenStatus;
-import backend.wal.onboard.domain.common.WalCategoryType;
-import backend.wal.onboard.domain.common.WalTimeType;
+import backend.wal.wal.common.domain.WalCategoryType;
+import backend.wal.wal.common.domain.WalTimeType;
 import backend.wal.reservation.domain.aggregate.ShowStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
