+++ src/test/java/backend/wal/onboarding/domain/entity/WalTimeTypeTest.java
@@ -1,6 +1,6 @@
 package backend.wal.onboarding.domain.entity;
 
-import backend.wal.onboard.domain.common.WalTimeType;
+import backend.wal.wal.common.domain.WalTimeType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
