+++ src/test/java/backend/wal/wal/repository/ItemRepositoryTest.java
@@ -1,10 +1,10 @@
 package backend.wal.wal.repository;
 
 import backend.wal.config.JpaRepositoryTestConfig;
-import backend.wal.onboard.domain.nextwal.aggregate.Category;
-import backend.wal.onboard.domain.nextwal.aggregate.Item;
-import backend.wal.onboard.domain.nextwal.repository.CategoryRepository;
-import backend.wal.onboard.domain.nextwal.repository.ItemRepository;
+import backend.wal.wal.nextwal.domain.aggregate.Category;
+import backend.wal.wal.nextwal.domain.aggregate.Item;
+import backend.wal.wal.nextwal.domain.repository.CategoryRepository;
+import backend.wal.wal.nextwal.domain.repository.ItemRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
